fix imports after move to scippneutron
